+++ toshi_hazard_store/query.py
@@ -1,2 +1,2 @@
 # """Queries for saving and retrieving openquake hazard results with convenience."""
-from toshi_hazard_store.gridded_hazard_query import get_gridded_hazard  # noqa
+from toshi_hazard_store.gridded_hazard_query import get_gridded_hazard, get_one_gridded_hazard  # noqa